Start of AssemblyPipelinePageIT testing
